review: move context and provider to distinct directory
